requirements.txt
Flask==3.1.2
typer==0.9.0
python-dateutil==2.9.0.post0
colorama==0.4.6
setuptools>=75.0.0
